refactor parse module to separate language files
