+++ src/parse/index.ts
@@ -1,0 +1,2 @@
+export { parseTs } from './typescript';
+export { parseVue } from './vue';
