+++ src/parse/typescript.ts
@@ -1,11 +1,10 @@
-import * as vueParser from 'vue-eslint-parser';
-import * as recast from 'recast';
-import { visit } from 'ast-types';
 
+import { visit } from 'ast-types';
 import * as babelParser from '@babel/parser';
-import { findAll } from './ast-helpers';
-import { VDocumentFragment } from './ast';
-import { VueProgram } from './types';
+import { File } from '@babel/types';
+import * as recast from 'recast';
+import { VueProgram } from '../types';
+
 
 const babelOptions = (isJsx: boolean): babelParser.ParserOptions => ({
   sourceType: 'module',
@@ -58,8 +57,9 @@ const babelOptions = (isJsx: boolean): babelParser.ParserOptions => ({
     ...isJsx ? ['jsx'] as any[] : [],
   ],
 });
-const tsParser = (isJsx: boolean) => ({
-  parse: (code: string) => babelParser.parse(code, babelOptions(isJsx)),
+
+export const tsParser = (isJsx: boolean) => ({
+  parse: (code: string): babelParser.ParseResult<File> => babelParser.parse(code, babelOptions(isJsx)),
   parseForESLint: (code: string) => {
     const res = babelParser.parse(code, babelOptions(isJsx));
 
@@ -92,50 +92,6 @@ const tsParser = (isJsx: boolean) => ({
   },
 });
 
-/**
- * Parse Vue code
- * @param code Source code
- * @returns SFC AST and Script AST
- */
-export function parseVue(code: string) {
-  if (!code.includes('<template')) {
-    // hack: if no <template> is present, templateBody will be null
-    // and we cannot access the VDocumentFragment
-    code += '\n<template></template>';
-  }
-
-  const sfcAST = vueParser.parse(code, {
-    parser: tsParser(true),
-    sourceType: 'module',
-  });
-
-  const scripts = findAll(sfcAST.templateBody!.parent as VDocumentFragment, {
-    type: 'VElement',
-    name: 'script',
-  }) as vueParser.AST.VElement[];
-
-  const scriptASTs = scripts.map((el) => {
-    // hack: make the source locations line up properly
-    const blankLines = '\n'.repeat(el.loc.start.line - 1);
-    const start = el.children[0]?.range[0];
-    const end = el.children[0]?.range[1];
-
-    const isJsx = el.startTag.attributes.some((attr) => !attr.directive && attr.key.rawName === 'lang' && attr.value && ['jsx', 'tsx'].includes(attr.value.value));
-
-    const ast = recast.parse(`/* METAMORPH_START */${blankLines}${code.slice(start, end)}`, {
-      parser: tsParser(isJsx),
-    }).program as VueProgram;
-
-    ast.isScriptSetup = el.startTag.attributes.some((attr) => !attr.directive && attr.key.rawName === 'setup');
-
-    return ast;
-  });
-
-  return {
-    sfcAST,
-    scriptASTs,
-  };
-}
 
 /**
  * Parse JS or TS code
